% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_parents_TGV.R
\name{calc_parents_TGV}
\alias{calc_parents_TGV}
\title{Cacluate Parents Total Genetic Value}
\usage{
calc_parents_TGV(parents, map.info, A, a, dom.coeff, usesnpeffect = F,
  snp.effects = NULL, save = F, prefix = NULL)
}
\arguments{
\item{parents}{The object that is returned from the create_parents function}

\item{map.info}{The object returned from create_map function}

\item{A}{Value assigned to the Major SNPQTL allele}

\item{a}{Value assigned to the Minor SNPQTL allele}

\item{dom.coeff}{The dominance coeffcient used for SNPQTLs}

\item{usesnpeffect}{logical.  Set to TRUE if specificying user provided snp effects. Default: FALSE}

\item{snp.effects}{The vector of SNP effects to use. Only use if usesnpeffect = TRUE.}

\item{save}{logical. Saves the output of genetic map (Default: FALSE)}

\item{prefix}{Name prefix to add to ouptut if save = TRUE. Default does not write output}
}
\description{
This function esimates the total genetic value of parents produced from the create_parents function
}
\examples{
parent.TGV <- calc_parents_TGV(parents = the.parents, map.info = the.map, A = 1, a = -100, dom.coeff = 1)
}
\keyword{genetic}
\keyword{parent}
\keyword{value}
