% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_progeny_TGV.R
\name{calc_progeny_TGV}
\alias{calc_progeny_TGV}
\title{Cacluate Progeny Total Genetic Value}
\usage{
calc_progeny_TGV(prog.info, map.info, cross.design, A, a, dom.coeff, save = F,
  prefix = NULL)
}
\arguments{
\item{prog.info}{The object that is returned from the make_crosses()}

\item{map.info}{The object returned from create_map()}

\item{cross.design}{Object returned from create_cross_design()}

\item{A}{Value assigned to the Major SNPQTL allele}

\item{a}{Value assigned to the Minor SNPQTL allele}

\item{dom.coeff}{The dominance coeffcient used for SNPQTLs}

\item{save}{logical. Saves the output of genetic map (Default: FALSE)}

\item{prefix}{Name prefix to add to ouptut if save = TRUE. Default does not write output}

\item{usesnpeffect}{logical.  Set to TRUE if specificying user provided snp effects. Default: FALSE}

\item{snp.effects}{The vector of SNP effects to use. Only use if usesnpeffect = TRUE.}
}
\description{
This function esimates the total genetic value of parents produced from the create_progeny function
}
\examples{
progeny1.TGV <- calc_progeny_TGV(prog.info = progeny1, map.info = the.map, crossdesign = cross.file,
A = 1, a = -100, dom.coeff = 1)
}
\keyword{genetic}
\keyword{progeny}
\keyword{value}
