% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cross_design.R
\name{create_cross_design}
\alias{create_cross_design}
\title{Simulate the creation of a cross design object}
\usage{
create_cross_design(parent.info, prog.percross, generation,
  mating.design = "AssortiveMating", coancest.thresh = T,
  use.par.marker.effects = F, use.op.par.phenos = F, cross.file = NULL,
  save = F, parent.phenos = NULL)
}
\arguments{
\item{parent.info}{The parent object from which contains individuals to make crosses from. This should come from either the create_parents() or extract_selections()}

\item{prog.percross}{Number of progeny that should be made per cross}

\item{generation}{Number that specifies which generation}

\item{mating.design}{Specify the type of mating design to be used in making crosses: "AssortiveMating" (Default), "Self", "RandomMating","SP","MP", or "cross.file.input". (See details for description)}

\item{coancest.thresh}{logical. Should a coancestry threshold be used as to minimize making related crosses?}

\item{use.par.marker.effects}{logical.  Set to TRUE if using the parents marker effects}

\item{use.op.par.phenos}{logical. Set to TRUE if passing parents created from OP_crossing()}

\item{cross.file}{User provided cross.file; Should be used if mating.design = "cross.design.input"}

\item{save}{logical. Saves the output of genetic map (Default: FALSE)}

\item{parent.phenos}{Object returned from sim_parent_phenos()}
}
\description{
This function creates the cross design object which can be generated from parents or the previous generation selections
}
\examples{
cross.file  <- create_cross_design(mapinfo = the.map,parent.info = op.families,prog.percross = 60,
generation = 1, use.op.par.phenos = T)
}
\keyword{create}
\keyword{cross}
\keyword{design}
