% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_genetic_map}
\alias{create_genetic_map}
\title{Simulate Genetic Map}
\usage{
create_genetic_map(num.chromos, map.length, num.markers, total.QTL, num.SNPQTL,
  distribute.loci = NULL, marker.distribution = NULL,
  map.dist = "haldane", chromosome.size.range = 0.2, total.loci = NULL,
  signif.digits = 2, marker.maf = NULL, seed = F, save = F)
}
\arguments{
\item{num.chromos}{The number of chromosomes to simulate}

\item{map.length}{The genetic map length in centi-morgans}

\item{num.markers}{Number of markers that can be used for genomic selection or estimated relatedness}

\item{total.QTL}{The total number of QTL to place on the genetic map}

\item{num.SNPQTL}{The number of total.QTL which will be SNPs}

\item{map.dist}{Type of distance function to use in calculating recombination frequences.  Options are: "haldane" & "kosambi".  (Default = "haldane)}

\item{chromosome.size.range}{The chromosome size range from the mean length of chromosomes (Default: 0.2)}

\item{total.loci}{The total number of loci to be placed on genetic map}

\item{signif.digits}{The number of significant digits to use (Default: 2)}

\item{seed}{logical. Set seed so that genetic map is the same every time (Default: FALSE)}

\item{save}{logical. Saves the output of genetic map (Default: FALSE)}

\item{het.markers}{logical. Should all markers used be fully heterozygous? (Default: FALSE)}
}
\description{
This function simulates the creation of a genetic map which mimics the genetic architecture of user defined inputs
}
\examples{
create_genetic_map(num.chromos = 12, map.length = 1800, num.markers = 120, total.loci = signif((120+2500)/12,digits = 2)*12, total.QTL = 2500, num.SNPQTL = 1960)
}
\keyword{cat}
