% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parents.R
\name{create_parents}
\alias{create_parents}
\title{Create parents for base population}
\usage{
create_parents(map.info, num.parents, max.delt.allele,
  heterozygous.markers = F, inbred.parents = F, QTL.sd = 0.25, save = F)
}
\arguments{
\item{map.info}{The object returned from create_map function}

\item{num.parents}{Number of parents that should be generated}

\item{max.delt.allele}{The Maximum number of deleterious alleles any single parent can have}

\item{heterozygous.markers}{logical. Should the markers for all parents be heterozygous?  Default: FALSE}

\item{QTL.sd}{A number providing the standard deviation of random QTL effects generated for the parents. Default: 0.25}

\item{save}{logical. Saves the output of genetic map (Default: FALSE)}
}
\description{
This function creates the base population that can be used for testing further mating and selection strategies
}
\examples{
the.parents <- create_parents(map.info = the.map, num.parents = 96, max.delt.allele = 14)
}
\keyword{create}
\keyword{parents}
