% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_selections.R
\name{extract_selections}
\alias{extract_selections}
\title{Extract selections from current generation}
\usage{
extract_selections(num.cores = NULL, map.info, cross.design, past.tgv,
  past.phenos, parent.info, progeny.info, progeny.TGV, progeny.phenos,
  selection.strategy, among.family.selection = NULL,
  within.family.selection = NULL, num.selections.within.family = 1,
  num.selections.among.family = NULL, relationship.matrix.type = "pedigree",
  prefix = NULL, rep.num = NULL, reduced = F, weighted = F)
}
\arguments{
\item{num.cores}{Number of cores used to create marker based relatedness matrix}

\item{map.info}{Object returned from create_genetic_map()}

\item{cross.design}{Object returned from create_cross_design()}

\item{past.tgv}{Object returned from previous generation}

\item{past.phenos}{Object returned from previous generation}

\item{parent.info}{Object returned from previous generation;}

\item{progeny.info}{Object returned from make_crosses()}

\item{progeny.TGV}{Object returned from calc_progeny_TGV()}

\item{progeny.phenos}{Object returned from sim_progeny_phenos()}

\item{selection.strategy}{Type of selection strategy to use. Currently either: "among.family" or "within.family". Default: "within.familiy"}

\item{among.family.selection}{Type of selection strategy to use for among family selection: "ABLUP", "GBLUP", "Phenotype"}

\item{within.family.selection}{Type of selection strategy to use for within family selection: "ABLUP", "GBLUP", "Phenotype"}

\item{num.selections.within.family}{The number of selections that should be made within each family}

\item{num.selections.among.family}{The number of selections that should be made among all families}

\item{relationship.matrix.type}{The type of relationship matrix to estimate for selections: "markers" or "pedigree"}

\item{prefix}{A prefix to add to the file if writing the file out}

\item{rep.num}{The rep number of the current generation}

\item{reduced}{logical. Set to TRUE if selecting less individuals than are in the total population}

\item{weighted}{logical. Should relationship matrix be weighted by A matrix prior to BLUP estimation.}

\item{save}{logical. Saves the output of genetic map (Default: FALSE)}
}
\description{
This function extracts selections from the current generation
}
\examples{
progeny1.extractions <- extract_selections(num.cores = 3,map.info = the.map,cross.design = cross.file,progeny.info = progeny1,
parent.info = op.families,progeny.TGV = progeny1.TGV,progeny.phenos = progeny1.PHENOS,num.selections.among.family = 64,reduced = T,
relationship.matrix.type = "markers",among.family.selection = "GBLUP",within.family.selection = "NO",selection.strategy = "within.family",
past.tgv = parent.TGV,past.phenos = parent.PHENOS)
}
\keyword{extract}
\keyword{selections}
