% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_crosses.R
\name{make_crosses}
\alias{make_crosses}
\title{Make crosses}
\usage{
make_crosses(parent.info, map.info, cross.design, run.parallel = F,
  num.cores = NULL)
}
\arguments{
\item{parent.info}{Object returned from create_parents()}

\item{map.info}{Object returned from create_genetic_map()}

\item{cross.design}{Object returned from create_cross_design()}

\item{run.parallel}{logical. Set TRUE to run function in parallel}

\item{num.cores}{The number of cores to use if running in parallel}
}
\description{
This function makes crosses among parents in a format specified by the cross design file
}
\examples{
progeny1 <- make_crosses(parent.info = op.families,map.info = the.map,cross.design = cross.file,run.parallel = T,num.cores = 3)
}
\keyword{among}
\keyword{create}
\keyword{crosses}
\keyword{parents}
