% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_testing.R
\name{OP_testing}
\alias{OP_testing}
\title{Generate Open Pollinated Families}
\usage{
OP_testing(map.info, parent.info, parent.phenos, parents.TGV, num.select,
  cross.prog = 1, dom.coeff, A, a, h2, E.var = NULL, run.parallel = F,
  n.cores = NULL)
}
\arguments{
\item{map.info}{Object returned from create_genetic_map()}

\item{parent.info}{Object returned from create_parents()}

\item{parent.phenos}{Object returned from sim_parents_phenos()}

\item{parents.TGV}{Object returned from calc_parents_TGV()}

\item{num.select}{The number of parents to select from the OP cross design}

\item{cross.prog}{The number of crosses to be made between each founder parent}

\item{dom.coeff}{The dominance cofficient used in estimating genetic values}

\item{A}{The value assigned to the major allele}

\item{a}{The value assigned to the minor allele}

\item{h2}{Heritability to be used in estimating phenotypes}

\item{run.parallel}{logical. Set TRUE to run in parallel.}

\item{n.cores}{The number of cores which will be used if ran in parallel}
}
\description{
This function simulates the creation of an Open Pollinated design and can make selections of which Parents perform the best
}
\examples{
op.families <- OP_testing( map.info = the.map, parent.info = the.parents, parent.phenos = parent.PHENOS,
parents.TGV = parent.TGV, cross.prog = 1, num.select = 64, dom.coeff = 1, A = 1, a = -100, h2 = .3, 
run.parallel = T,n.cores = 3)
}
\keyword{Generate}
\keyword{OP}
\keyword{families}
