% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_parents_phenos.R
\name{sim_parents_phenos}
\alias{sim_parents_phenos}
\title{Simulate parental phenotypes}
\usage{
sim_parents_phenos(parents.TGV, h2, E.var = NULL, save = F, folder = NULL)
}
\arguments{
\item{parents.TGV}{Object returned from calc_parent_TGV()}

\item{h2}{Heritability used to simulate phenotypes}

\item{E.var}{Environmental variance that can be used instead of heritability}

\item{save}{logical.  Set to TRUE to save output.  Default: FALSE}

\item{folder}{Folder to output results to. Only matters if save=T}
}
\description{
This function generates parental phenotypes and is used after creating the founder population
}
\examples{
parent.PHENOS <- sim_parents_phenos(parents.TGV = parent.TGV,h2 = .3)
}
\keyword{phenos}
\keyword{progeny}
\keyword{sim}
