% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_progeny_phenos.R
\name{sim_progeny_phenos}
\alias{sim_progeny_phenos}
\title{Simulate progeny phenos}
\usage{
sim_progeny_phenos(progeny.TGV, h2, cross.design, E.var = NULL, save = F,
  folder = NULL)
}
\arguments{
\item{progeny.TGV}{Object returned from calc_progeny_TGV()}

\item{h2}{Heritability used to simulate phenotypes}

\item{cross.design}{Object returned from create_cross_design()}

\item{E.var}{Environmental variance that can be used instead of heritability}

\item{save}{logical.  Set to TRUE to save output.  Default: FALSE}

\item{folder}{Folder to output results to. Only matters if save=T}
}
\description{
This function simulates progeny phenotypes
}
\examples{
progeny1.PHENOS <- sim_progeny_phenos(progeny.TGV = progeny1.TGV,h2 = .3,cross.design = cross.file)
}
\keyword{phenos}
\keyword{progeny}
\keyword{sim}
